Add playerJump

+++ src/main/java/playing/entities/player/playerModules/PlayerHitBox.java
@@ -45,12 +45,12 @@ public class PlayerHitBox extends PlayerModule implements PlayingUpdateInterface
         return hitBox;
     }
 
-    public void updateHitBoxX(int xSpeed){
+    public void updateHitBoxX(double xSpeed){
         hitBox.x += xSpeed;
         playerModuleManager.setPlayerX(hitBox.x);
     }
 
-    public void updateHitBoxY(int ySpeed){
+    public void updateHitBoxY(double ySpeed){
         hitBox.y += ySpeed;
         playerModuleManager.setPlayerY(hitBox.y);
     }
